serialized tests and reorganized files
